DKMS: Update nvidia*-use-ARCH.o_binary.patch for arm64
nvidia-kernel-dkms is packaged for arm64, with the binary object files
renamed to arch-specific names, but without the patch to use the arch-
specific name in the arm64 case. This causes the final link to fail
when DKMS builds the nvidia.ko and nvidia-modeset.ko kernel modules,
as Kbuild is unable to resolve the nv*-kernel.o files for linking.

Update the nvidia*-use-ARCH.o_binary.patch patches to add a line
defining the relevant binary object file names to the arm64 variant
when CONFIG_ARM64 is set on the target kernel.